🏗 API changes, removes excess "shell" components
